CLN: refactor auth logic to its own module.
Uses new-style tests for auth-related tests.
